+++ lib/src/theme/control_button_bar_theme.dart
@@ -47,22 +47,25 @@ class ControlButtonBarTheme extends InheritedWidget {
 /// [ControlButtonBar] widgets when used with [ControlButtonBarTheme] or with the overall
 /// {@endtemplate}
 class ControlButtonBarThemeData with Diagnosticable {
-  factory ControlButtonBarThemeData(
-    BuildContext context, {
+  factory ControlButtonBarThemeData({
     double? height,
     double? buttonSpacing,
     BoxDecoration? decoration,
     ControlButtonThemeData? controlButtonTheme,
+    Color? backgroundColor,
     EdgeInsetsGeometry? margin,
     EdgeInsetsGeometry? padding,
   }) {
-    controlButtonTheme ??= ControlButtonThemeData.light(context);
+    controlButtonTheme ??= ControlButtonThemeData.light();
     height ??= 50;
     margin ??= EdgeInsets.symmetric(horizontal: 10, vertical: 10);
     padding ??= EdgeInsets.symmetric(horizontal: 10);
     buttonSpacing ??= 0;
+    // backgroundColor ??= decoration == null ? Colors.white : null;
+    if (decoration == null) {
+      backgroundColor ??= Colors.white;
+    }
     decoration ??= const BoxDecoration(
-      color: Colors.white,
       borderRadius: BorderRadius.all(Radius.circular(25)),
       boxShadow: <BoxShadow>[
         BoxShadow(
@@ -73,6 +76,7 @@ class ControlButtonBarThemeData with Diagnosticable {
       ],
     );
     return ControlButtonBarThemeData.raw(
+      backgroundColor: backgroundColor,
       controlButtonTheme: controlButtonTheme,
       buttonSpacing: buttonSpacing,
       controlContainerDecoration: decoration,
@@ -89,8 +93,8 @@ class ControlButtonBarThemeData with Diagnosticable {
     this.height = 50,
     this.margin = const EdgeInsets.symmetric(horizontal: 10, vertical: 10),
     this.padding = const EdgeInsets.symmetric(horizontal: 10),
+    this.backgroundColor,
     this.controlContainerDecoration = const BoxDecoration(
-      color: Colors.white,
       borderRadius: BorderRadius.all(Radius.circular(25)),
       boxShadow: <BoxShadow>[
         BoxShadow(
@@ -102,14 +106,13 @@ class ControlButtonBarThemeData with Diagnosticable {
     ),
   });
 
-  factory ControlButtonBarThemeData.light(BuildContext context) =>
-      ControlButtonBarThemeData(
-        context,
+  factory ControlButtonBarThemeData.light() => ControlButtonBarThemeData(
         margin: EdgeInsets.symmetric(horizontal: 10, vertical: 10),
         padding: EdgeInsets.zero,
         buttonSpacing: 0,
         height: 50,
-        controlButtonTheme: ControlButtonThemeData.light(context),
+        controlButtonTheme: ControlButtonThemeData.light(),
+        backgroundColor: Colors.white,
         decoration: const BoxDecoration(
           color: Colors.white,
           borderRadius: BorderRadius.all(Radius.circular(25)),
@@ -121,6 +124,17 @@ class ControlButtonBarThemeData with Diagnosticable {
             )
           ],
         ),
+      );
+
+  factory ControlButtonBarThemeData.dark(BuildContext context) =>
+      ControlButtonBarThemeData(
+        buttonSpacing: 20,
+        backgroundColor: Color(0xff19355D),
+        decoration: BoxDecoration(
+          color: Color(0xff19355D),
+          borderRadius: BorderRadius.circular(50),
+        ),
+        controlButtonTheme: ControlButtonThemeData.dark(),
       );
 
   /// {@template control_container_decoration}
@@ -155,6 +169,9 @@ class ControlButtonBarThemeData with Diagnosticable {
   /// Height of the control button bar
   final double height;
 
+  /// Background of the control button bar
+  final Color? backgroundColor;
+
   @override
   bool operator ==(Object other) =>
       identical(this, other) ||
@@ -162,6 +179,7 @@ class ControlButtonBarThemeData with Diagnosticable {
           runtimeType == other.runtimeType &&
           controlContainerDecoration == other.controlContainerDecoration &&
           buttonSpacing == other.buttonSpacing &&
+          backgroundColor == other.backgroundColor &&
           controlButtonTheme == other.controlButtonTheme &&
           margin == other.margin &&
           padding == other.padding &&
@@ -171,6 +189,7 @@ class ControlButtonBarThemeData with Diagnosticable {
   int get hashCode =>
       controlContainerDecoration.hashCode ^
       buttonSpacing.hashCode ^
+      backgroundColor.hashCode ^
       controlButtonTheme.hashCode ^
       controlContainerDecoration.hashCode ^
       margin.hashCode ^
@@ -182,6 +201,8 @@ class ControlButtonBarThemeData with Diagnosticable {
     super.debugFillProperties(properties);
     properties.add(DiagnosticsProperty<BoxDecoration>(
         'controlContainerDecoration', controlContainerDecoration));
+    properties
+        .add(DiagnosticsProperty<Color>('backgroundColor', backgroundColor));
     properties.add(DiagnosticsProperty<ControlButtonThemeData>(
         'controlButtonTheme', controlButtonTheme));
     properties.add(DiagnosticsProperty<double>('buttonSpacing', buttonSpacing));