+++ CHANGELOG.md
@@ -33,6 +33,7 @@
 - Add `ControlButtonTheme` for control button theme
 - Add `controlButtons` option to display/hide control buttons (All, Reset)
 - Add `FilterListDelegate.show` delegate to search/filter data in new screen
+- Add `FilterListDelegateTheme` for filter list delegate theme
 
 
 ## [1.0.0] 21 May 2021